requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-fab
apache-airflow-providers-http==5.4.0
apache-airflow-providers-amazon==9.20.0
apache-airflow-providers-mysql==6.4.1
apache-airflow-providers-postgres==6.5.1
numpy==1.26.4
pandas==2.1.4
polars[database]==1.37.1
#adbc-driver-manager 
#adbc-driver-postgresql